fix doctests
